requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
langgraph
langchain
langchain-core
langchain-google-genai
pydantic
